pyproject.toml
[project]
name = "SeqLP"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Nils Hofmann", email = "nilhof01@gmail.com"},
    {name = "nilshof01", email = "n.hofmann.99@web.de"},
]
dependencies = [
    "pytest>=7.1.2",
    "requests>=2.31.0",
    "transformers>=4.39.3",
    "pandas>=2.2.1",
    "scikit-learn>=1.4.1.post1",
    "editdistance==0.6.2",
    "torch>=2.2.2",
    "biopython",
    "matplotlib",
    "seaborn"
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
package-dir = "src"
distribution = true

[tool.pdm.dev-dependencies]
test = [
    "pytest>=7.1.2",
]

[tool.pdm.scripts]
test = "python -m pytest -s tests"